ai_viewer/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
pydantic>=2.0.0
